下载bug解决

+++ data-process-crawler-download/src/main/java/com/qjq/crawler/download/service/DownLoadService.java
@@ -4,7 +4,7 @@ import org.data.process.mqmodel.DownLoadMessage;
 
 public interface DownLoadService {
 
-    public void addSeed(String url, String jobId, Integer deep, Integer sleep);
+    public void addSeed(String url, String jobId, Integer deep, Long sleep);
 
     public void schudel(DownLoadMessage message);
